feat: rename hooks in kebab casing for consistency
